+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Introduce `logjam.appender.default-event-size` configuration option.
 
+* Event filters: rename `:loggers` to `:loggers-allowlist`
+ * `:loggers` will remain supported.
+
 ## 0.2.0 (2024-01-04)
 
 * [#8](https://github.com/clojure-emacs/logjam/issues/8): Introduce Timbre compatibility.